Try to save memory on storage of keys by using Arc in the HashMap's key
If size_of::<Arc<K>> (which is 8 on my machine) is smaller than the mean
sizeof::<K> (which depends on your data set) then this saves us memory space by
allowing us to only store the Key once and store a bunch of reference counted
pointers to it.

If incrementing an Arc is faster than copying the bytes then it will also be
faster. I think that this is unlikely but it could be true for very large keys.